Add dialogs and metadata modules for Substance Painter integration
- Introduced `dialogs.py` for asset selection and project creation
  workflows.
- Added `metadata.py` to handle asset metadata in Substance Painter
  projects.
- Updated `__init__.py` to include new modules in the package.
- Refactored `ui.py` to use the new `dialogs` and `metadata` modules.
