+++ pipeline/pipe/sp/__init__.py
@@ -1,10 +1,12 @@
 # import substance painter-specific scripts
-from . import assetfile, channels, export, local, reload
+from . import assetfile, channels, dialogs, export, local, metadata, reload
 
 __all__ = [
     "assetfile",
     "channels",
+    "dialogs",
     "export",
     "local",
+    "metadata",
     "reload",
 ]
